refactor: DB 컬럼 매핑 동기화 및 필드명 정리
- Cocktail: methodRaw, imageUrl @Transient -> 실제 컬럼 매핑
- Cocktail: proofInsideBracketProof -> proofInsideBracket (컬럼명 proof_inside_bracket)
- Ingredient: tier 타입 Long -> Integer

+++ src/main/java/com/cock/cocktail/domain/Ingredient.java
@@ -41,7 +41,7 @@ public class Ingredient {
 
     // 재료별 중요도 ( 자주 쓰이는 재료 기준 )
     @Column(name = "tier")
-    private Long tier;
+    private Integer tier;
 
     @Column(name = "frequency")
     private Long frequency;
